+++ changelog.md
@@ -5,6 +5,7 @@
   - Added `integrity` values to external assets (issue#10)
   - Move mapbox links into own sub-view (issue#11)
   - Updated mapbox version (issue#12)
+  - Massive refactor of webmention code, allowing for re-parse command (issue#8)
 
 ## Version 0.0.10 (2016-09-10)
   - Add an artisan command for sensiolab’s security check